Move AnyWindowVM to a separate file
